+++ src/variables/data.jl
@@ -48,12 +48,10 @@ isconst(::AbstractArray{<:DataVariable})  = false
 get_stream_of_marginals(datavar::DataVariable) = datavar.marginal
 get_stream_of_predictions(datavar::DataVariable) = datavar.prediction
 
-set_stream_of_marginals!(datavar::DataVariable, stream) = connect!(
-    datavar.marginal, stream
-)
-set_stream_of_predictions!(datavar::DataVariable, stream) = connect!(
-    datavar.prediction, stream
-)
+set_stream_of_marginals!(datavar::DataVariable, stream) =
+    connect!(datavar.marginal, stream)
+set_stream_of_predictions!(datavar::DataVariable, stream) =
+    connect!(datavar.prediction, stream)
 
 function create_new_stream_of_inbound_messages!(datavar::DataVariable)
     new_stream_of_inbound_messages = MessageObservable(AbstractMessage)
@@ -87,9 +85,8 @@ struct DataVariableActivationOptions
     args
 end
 
-DataVariableActivationOptions() = DataVariableActivationOptions(
-    false, false, nothing, nothing
-)
+DataVariableActivationOptions() =
+    DataVariableActivationOptions(false, false, nothing, nothing)
 
 """
     ReactiveMP.activate!(datavar::DataVariable, options::DataVariableActivationOptions)
@@ -150,9 +147,8 @@ end
 
 __link_getmarginal(constant) = of(Marginal(PointMass(constant), true, false))
 __link_getmarginal(l::AbstractVariable) = get_stream_of_marginals(l)
-__link_getmarginal(l::AbstractArray{<:AbstractVariable}) = collectLatest(
-    map(get_stream_of_marginals, l)
-)
+__link_getmarginal(l::AbstractArray{<:AbstractVariable}) =
+    collectLatest(map(get_stream_of_marginals, l))
 
 __apply_link(f::F, args) where {F} = __apply_link(f, getdata.(args))
 __apply_link(f::F, args::NTuple{N, PointMass}) where {F, N} = f(mean.(args)...)
@@ -165,9 +161,8 @@ Provides a new observation to a [`ReactiveMP.DataVariable`](@ref) (or an array o
 The `data` is wrapped in a `PointMass` distribution and pushed as a new message.
 Pass `missing` to indicate that the observation is not available.
 """
-new_observation!(datavar::DataVariable, data) = new_observation!(
-    datavar, PointMass(data)
-)
+new_observation!(datavar::DataVariable, data) =
+    new_observation!(datavar, PointMass(data))
 new_observation!(datavar::DataVariable, data::PointMass) = next!(datavar.messageout, Message(data, false, false))
 new_observation!(datavar::DataVariable, ::Missing)       = next!(datavar.messageout, Message(missing, false, false))